Added is_file checks around unlinks to prevent php warnings from causing header issues #bing existential question

+++ public/include/class.fulltext_index_solr_csv.php
@@ -383,7 +383,7 @@ class FulltextIndex_Solr_CSV extends FulltextIndex
 			}
 
 			// Dont delete csv if there is an error
-			if($uploaded == true) {
+			if($uploaded == true && is_file($tmpfname)) {
 				unlink($tmpfname);
 			}
 
@@ -567,7 +567,7 @@ class FulltextIndex_Solr_CSV extends FulltextIndex
 	 * @param string $dsID
 	 * @return plaintext of datastream, null on error
 	 */
-	public function getCachedContent($pids) 
+	public function getCachedContent($pids)
 	{
 		$log = FezLog::get();
 		$db = DB_API::get();
